// src/api/generated/model/userCreate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * LeadRadar API
 * OpenAPI spec version: 0.1.0
 */
import type { UserCreateRole } from './userCreateRole';

export interface UserCreate {
  email: string;
  password: string;
  full_name?: string | null;
  role?: UserCreateRole;
  org_id?: string | null;
}