// src/api/generated/model/userUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * LeadRadar API
 * OpenAPI spec version: 0.1.0
 */
import type { UserUpdateRole } from './userUpdateRole';

export interface UserUpdate {
  full_name?: string | null;
  role?: UserUpdateRole;
  is_active?: boolean | null;
  password?: string | null;
}